chore: webhook and pipeline types
