+++ apps/web/app/(app)/environments/[environmentId]/integrations/webhooks/components/AddWebhookModal.tsx
@@ -6,6 +6,7 @@ import { Button } from "@/modules/ui/components/button";
 import { Input } from "@/modules/ui/components/input";
 import { Label } from "@/modules/ui/components/label";
 import { Modal } from "@/modules/ui/components/modal";
+import { PipelineTriggers } from "@prisma/client";
 import clsx from "clsx";
 import { Webhook } from "lucide-react";
 import { useTranslations } from "next-intl";
@@ -13,10 +14,9 @@ import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
 import toast from "react-hot-toast";
-import { TPipelineTrigger } from "@formbricks/types/pipelines";
 import { TSurvey } from "@formbricks/types/surveys/types";
-import { TWebhookInput } from "@formbricks/types/webhooks";
 import { createWebhookAction, testEndpointAction } from "../actions";
+import { TWebhookInput } from "../types/webhooks";
 
 interface AddWebhookModalProps {
   environmentId: string;
@@ -37,7 +37,7 @@ export const AddWebhookModal = ({ environmentId, surveys, open, setOpen }: AddWe
   const [testEndpointInput, setTestEndpointInput] = useState("");
   const [hittingEndpoint, setHittingEndpoint] = useState<boolean>(false);
   const [endpointAccessible, setEndpointAccessible] = useState<boolean>();
-  const [selectedTriggers, setSelectedTriggers] = useState<TPipelineTrigger[]>([]);
+  const [selectedTriggers, setSelectedTriggers] = useState<PipelineTriggers[]>([]);
   const [selectedSurveys, setSelectedSurveys] = useState<string[]>([]);
   const [selectedAllSurveys, setSelectedAllSurveys] = useState(false);
   const [creatingWebhook, setCreatingWebhook] = useState(false);
@@ -88,7 +88,7 @@ export const AddWebhookModal = ({ environmentId, surveys, open, setOpen }: AddWe
     );
   };
 
-  const handleCheckboxChange = (selectedValue: TPipelineTrigger) => {
+  const handleCheckboxChange = (selectedValue: PipelineTriggers) => {
     setSelectedTriggers((prevValues) =>
       prevValues.includes(selectedValue)
         ? prevValues.filter((value) => value !== selectedValue)
