+++ apps/web/app/(app)/environments/[environmentId]/integrations/webhooks/components/WebhookDetailModal.tsx
@@ -1,15 +1,15 @@
 import { WebhookOverviewTab } from "@/app/(app)/environments/[environmentId]/integrations/webhooks/components/WebhookOverviewTab";
 import { WebhookSettingsTab } from "@/app/(app)/environments/[environmentId]/integrations/webhooks/components/WebhookSettingsTab";
 import { ModalWithTabs } from "@/modules/ui/components/modal-with-tabs";
-import { Webhook } from "lucide-react";
+import { Webhook } from "@prisma/client";
+import { WebhookIcon } from "lucide-react";
 import { useTranslations } from "next-intl";
 import { TSurvey } from "@formbricks/types/surveys/types";
-import { TWebhook } from "@formbricks/types/webhooks";
 
 interface WebhookModalProps {
   open: boolean;
   setOpen: (v: boolean) => void;
-  webhook: TWebhook;
+  webhook: Webhook;
   surveys: TSurvey[];
   isReadOnly: boolean;
 }
@@ -35,7 +35,7 @@ export const WebhookModal = ({ open, setOpen, webhook, surveys, isReadOnly }: We
         open={open}
         setOpen={setOpen}
         tabs={tabs}
-        icon={<Webhook />}
+        icon={<WebhookIcon />}
         label={webhook.name ? webhook.name : webhook.url}
         description={""}
       />
