+++ apps/web/app/(app)/environments/[environmentId]/integrations/webhooks/components/WebhookRowData.tsx
@@ -1,12 +1,12 @@
 import { Badge } from "@/modules/ui/components/badge";
+import { Webhook } from "@prisma/client";
 import { useTranslations } from "next-intl";
 import { timeSince } from "@formbricks/lib/time";
 import { capitalizeFirstLetter } from "@formbricks/lib/utils/strings";
 import { TSurvey } from "@formbricks/types/surveys/types";
 import { TUserLocale } from "@formbricks/types/user";
-import { TWebhook } from "@formbricks/types/webhooks";
 
-const renderSelectedSurveysText = (webhook: TWebhook, allSurveys: TSurvey[]) => {
+const renderSelectedSurveysText = (webhook: Webhook, allSurveys: TSurvey[]) => {
   if (webhook.surveyIds.length === 0) {
     const allSurveyNames = allSurveys.map((survey) => survey.name);
     return <p className="text-slate-400">{allSurveyNames.join(", ")}</p>;
@@ -19,7 +19,7 @@ const renderSelectedSurveysText = (webhook: TWebhook, allSurveys: TSurvey[]) => 
   }
 };
 
-const renderSelectedTriggersText = (webhook: TWebhook, t: (key: string) => string) => {
+const renderSelectedTriggersText = (webhook: Webhook, t: (key: string) => string) => {
   if (webhook.triggers.length === 0) {
     return <p className="text-slate-400">No Triggers</p>;
   } else {
@@ -58,7 +58,7 @@ export const WebhookRowData = ({
   surveys,
   locale,
 }: {
-  webhook: TWebhook;
+  webhook: Webhook;
   surveys: TSurvey[];
   locale: TUserLocale;
 }) => {
