+++ apps/web/app/api/v1/webhooks/types/webhooks.ts
@@ -1,0 +1,16 @@
+import { z } from "zod";
+import { ZWebhook } from "@formbricks/database/zod/webhooks";
+
+export const ZWebhookInput = ZWebhook.partial({
+  name: true,
+  source: true,
+  surveyIds: true,
+}).pick({
+  name: true,
+  source: true,
+  surveyIds: true,
+  triggers: true,
+  url: true,
+});
+
+export type TWebhookInput = z.infer<typeof ZWebhookInput>;
