+++ apps/web/app/lib/pipelines.ts
@@ -1,5 +1,5 @@
+import { TPipelineInput } from "@/app/lib/types/pipelines";
 import { CRON_SECRET, WEBAPP_URL } from "@formbricks/lib/constants";
-import { TPipelineInput } from "@formbricks/types/pipelines";
 
 export const sendToPipeline = async ({ event, surveyId, environmentId, response }: TPipelineInput) => {
   return fetch(`${WEBAPP_URL}/api/pipeline`, {
