+++ apps/web/lib/utils/services.ts
@@ -2,6 +2,7 @@
 
 import { contactCache } from "@/lib/cache/contact";
 import { teamCache } from "@/lib/cache/team";
+import { webhookCache } from "@/lib/cache/webhook";
 import { Prisma } from "@prisma/client";
 import { cache as reactCache } from "react";
 import { prisma } from "@formbricks/database";
@@ -18,7 +19,6 @@ import { responseNoteCache } from "@formbricks/lib/responseNote/cache";
 import { surveyCache } from "@formbricks/lib/survey/cache";
 import { tagCache } from "@formbricks/lib/tag/cache";
 import { validateInputs } from "@formbricks/lib/utils/validate";
-import { webhookCache } from "@formbricks/lib/webhook/cache";
 import { ZId, ZString } from "@formbricks/types/common";
 import { DatabaseError, InvalidInputError, ResourceNotFoundError } from "@formbricks/types/errors";
